322、商城业务-秒杀服务-秒杀流程

+++ gulimall-seckill/src/main/java/com/atguigu/gulimall/seckill/to/SecKillSkuRedisTo.java
@@ -35,7 +35,7 @@ public class SecKillSkuRedisTo {
     /**
      * 每人限购数量
      */
-    private BigDecimal seckillLimit;
+    private Integer seckillLimit;
     /**
      * 排序
      */
